Resolve import issues
